+++ empyrebase/services/auth.py
@@ -227,6 +227,21 @@ class Auth:
         raise_detailed_error(request_object)
         return request_object.json()
 
+    def change_email(self, id_token, new_email, return_secure_token=True):
+        """
+        https://firebase.google.com/docs/reference/rest/auth#section-change-email
+        """
+        request_ref = f"https://identitytoolkit.googleapis.com/v1/accounts:update?key={self.api_key}"
+        headers = {"content-type": "application/json; charset=UTF-8"}
+        data = json.dumps({
+            "idToken": id_token,
+            "email": new_email,
+            "returnSecureToken": return_secure_token
+        })
+        request_object = requests.post(request_ref, headers=headers, data=data)
+        raise_detailed_error(request_object)
+        return request_object.json()
+
     def change_password(self, id_token: str, password: str, return_secure_token: bool):
         """Change the password of a user